Handle improperly formed XML in launchd provider
Previously, we implemented CFPropertyList with the launchd service
provider.  In several cases, improperly formed XML (such as the kind
found in Apple's own /System/Library/LaunchDaemons/org.cups.cupsd.plist
will cause CFPropertyList's parsers to fail (notably libxml and
nokogiri).  Because we can't repair this behavior like we can with
improperly commented Doctype strings, we must output an error, skip the
file, and continue on.  This commit implements that behavior.