Use PBKDF2 Password Hashes on 10.8
Previously, if a machine had been upgraded from 10.7 to 10.8 and user
existed with a 10.7-style Salted-SHA512 hash, and Puppet tried to
enforce a 10.8-style PBKDF2 password hash, Puppet would throw an
error. With this code change, we remove the 10.7-style hash, create
a 10.8-style hash, and then move on with the Puppet run.