Fix issue with upgraded password values
A previous commit reverted the use of an instance variable for the
'converted_hash_plist' variable because it allowed for stale data to
populate a user's plist in a single Puppet run. This code accounts for
users that have a 10.7-style password hash to have a 10.8-style password
hash set for them by Puppet.  Puppet does this by deleting the
'SALTED-SHA512' key in the converted_hash_plist hash and instead adding
a 'SALTED-SHA512-PBKDF2' key.