Clear ds cache for password calls
Methods around setting the password on OS X are the ONLY methods that
cannot use dscl (because the only way to set it via dscl is by passing
a plaintext password - which is bad). Because of this, we have to change
the user's plist directly. DSCL has its own caching mechanism, which
means that every time we call dscl in this provider we're not directly
changing values on disk (instead, those calls are cached and written
to disk according to Apple's prioritization algorithms). When Puppet
needs to set the password property on OS X > 10.6, the provider has to
tell dscl to write its cache to disk before modifying the user's
plist. The 'dscacheutil -flushcache' command does this. Another issue
is how fast Puppet makes calls to dscl and how long it takes dscl to
enter those calls into its cache. We have to sleep for 2 seconds before
flushing the dscl cache to allow all dscl calls to get INTO the cache
first. This could be made faster (and avoid a sleep call) by finding
a way to enter calls into the dscl cache faster. A sleep time of 1
second would intermittantly require a second Puppet run to set
properties, so 2 seconds seems to be the minimum working value.