Check if dscl returns a value in dynamic getters
Previously, in the block of dynamically-generated getter methods,
the method didn't successfully determine if dscl returned an array
of hashes.  If a value was empty, the provider threw an error because
it tried to reference the first element in a nil value. This change
checks if a value is returned first, and then dereferences the first
value in the array.