Helper method for writing files to disk
Previously, the write_sha1_hash method would use File.open to
directly write files to disk. This is problematic to test
because it requires mocking File.open (which can lead to
unintended consequences in the event that File.open is
ACTUALLY necessary later). To remedy this, a helper method of
write_to_file has been created to facilitate easier mocking
in the unit tests, and this method is called from the
write_sha1_hash method.