Helper method for creating a user
Previously, the dscl binary was being called directly from
the create method. This required mocking dscl from the unit
tests. While this was successful, it's easier to wrap the
process in a helper method so that the process used to
create the user could be changed at a later time and we're
not tied to using dscl.