Helper method for merging dscl values
Previously, dscl was being called directly to merge
attribute values. This worked, but it created a direct
reliance on the dscl binary both in the code as well as
in the tests. This meant that the dscl command had to be
mocked out in all our tests involving changing attributes.
To decouple the provider and the tests from this behavior,
a helper method was created that makes the dscl calls. This
allows us to mock out the helper method in the tests, and
leaves the door open for refactoring in the future (in the
event that another method OTHER than directly calling dscl
proves to work better).