Update launchd test for 1.9.3
Previously, the spec tests for the launchd service provider would
fail on Ruby 1.9.3 because the mocked objects for the launchd_paths
and return_globbed_list_of_file_paths methods would return strings
and not arrays (and in Ruby 1.8.7, you could call .each on a string
and have it return the string object).  This commit fixes that by
returning an Array and not a String.