Rescue ArgumentError in rbCFPropertyList.rb
Previously, rbCFPropertyList.rb had instantiated an object of
Enumerable::Enumerator and passed it an argument of an empty Array.
This caused issues when running Puppet in Ruby 1.9.3 because it threw
a stacktrace giving an Argument Error (1 for 0) on the initialize
method of Enumerator.

The underlying issue is that Ruby 1.8.x has Enumerable::Enumerator
whereas Ruby 1.9.x has moved Enumerator into the global namespace.
It seems as if an Enumerator object has already been instantiated
within the Puppet codebase, so rescuing this ArgumentError should
be considered pretty 'brute force' at the moment.  This works for
now, however a better solution is ideal.